+++ checkpoint/CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - #v1 Make most V1 public concrete classes final.
+- Refactor `CheckpointLayout` splitting `load()` into `load_pytree()` and
+`load_checkpointables()` each with their own dedicated loading logic
 
 ## [0.11.32] - 2026-01-20